refactor(utils): extract TsconfigInfo dataclass to tsconfig_info.py
Move TsconfigInfo out of the monolithic tsconfig.py into its own module
so the dataclass definition is easy to locate. tsconfig.py re-exports it
for backward compatibility. Also fix two CR bugs: pass set(_seen) copies
to sibling extends branches so shared visited state doesn't cause false
cycle detection, and extend has_explicit_scoping() to return True when
exclude_patterns is non-empty (clearing it via temp tsconfig is equally
incorrect).
